+++ core/components/crosslinks/model/crosslinks/events/crosslinksonloadwebdocument.class.php
@@ -8,6 +8,21 @@ class CrosslinksOnLoadWebDocument extends CrosslinksPlugin
 {
     public function run()
     {
+        $contexts = $this->crosslinks->getOption('enabledContexts');
+        if ($contexts) {
+            $contexts = explode(',', $contexts);
+            if (!in_array($this->modx->context->key, $contexts)) {
+                return;
+            }
+        }
+        $templates = $this->crosslinks->getOption('enabledTemplates');
+        if ($templates) {
+            $templates = explode(',', $templates);
+            if (!in_array($this->modx->context->key, $templates)) {
+                return;
+            }
+        }
+
         $chunkName = $this->crosslinks->getOption('tpl');
 
         $content = $this->modx->resource->get('content');