+++ core/components/crosslinks/docs/changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.2.2] - TBA
+### Added
+- enabledContexts and enabledTemplates system setting
 ### Changed
 - Quote the the regular expression delimiter in the linktext